+++ resources/views/pages/banner.blade.php
@@ -4,9 +4,11 @@
 
         <div class="container maxw-lg space-y-1.5">
 
-            <div class="flex">
-                <h1 class="banner-title bg-blue-09">{{ $page->title }}</h1>
-            </div>
+            @unless($page->config['hide_title'])
+                <div class="flex">
+                    <h1 class="banner-title bg-blue-09">{{ $page->title }}</h1>
+                </div>
+            @endunless
 
             <div class="banner-text inline-flex">