+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.2
+- add warning message for advanced layouts
+- replace some database fields with config column
+
 ## 0.1.1
 - prevent slug update if one already exists